test: move etcd unit tests to their own package
